Reorganized functional testing
